Use mode instead of mean when calculating operator fixities
When deciding on the fixity of an operator, we assign a score based on some
heuristics. When an operator is used multiple times in an operator chain, we
used to calculate the mean of its scores to decide on the final fixity.
However, this was prone to outliers.

In an example like this:

  b & c .~ d
    & e
      %~ f
          g

The first use of & is surrounded by other values, so we assign it a higher
score (binds tighter). However the second occurence is at the beginning of a
line, and this has a lower score. In the end, when we average those two
values the result binds tighter than %~, which is not what we want.

This PR replaces the average method with a mode operation instead, so it is
less prone to outliers.

This PR fixes this specific issue, and does not cause any change on other
test cases, but it might cause some diffs in the wild.