Fix GHC panic on showOutputable
Closes #544.

It looks like when we call 'showSDocUnsafe' it uses `unsafeGlobalDynFlags`
wwhich the `settings` field is not initialised unless someone calls
`setUnsafeGlobalDynFlags`, `initGhcMonad` or `runGhc` and throws a panic.

I couldn't pinpoint the exact cause, but I suspect probably there was
a change either on Ormolu or ghc-lib-parser which removed a call to one
of the functions above before we print manualExts.

Anyway, we actually don't need to call `showSDocUnsafe` since we already
have a populated `DynFlags`. This commit removes the unsafe call and passes
the `DynFlags` we have to `showSDoc`.