Add line numbers to Ormolu Live editor
This uses a pure-CSS solution to add line numbers to the text editor
in Ormolu Live. It calculates the desired line number count from the
input, and then creates an equivalent number of `span` elements. A
CSS counter is then incremented for each span, and the content is set
to the counter's value.